+++ src/Components/Map.c
@@ -6,14 +6,14 @@ RenderData* LoadInformationMap(void)
     RenderData* tempData = (RenderData*)calloc(1,sizeof(RenderData));
     ///< Load Maps
     tempData->mapsData = (MapData*)calloc(2,sizeof(MapData));
-    tempData->mapsData[0].data = LoadMapTiles("assets/Maps/CGame_L1_BG_Layer.csv",&tempData->mapsData[0].width,&tempData->mapsData[0].height);
+    tempData->mapsData[0].data = LoadMapTiles(ASSETS"/Maps/CGame_L1_BG_Layer.csv",&tempData->mapsData[0].width,&tempData->mapsData[0].height);
     tempData->mapsData[0].size = (Rectangle){0,0,(float)tempData->mapsData[0].width * 16, (float)tempData->mapsData[0].height * 16};
-    tempData->mapsData[1].data = LoadMapTiles("assets/Maps/CGame_L1_UP_Layer.csv",&tempData->mapsData[1].width,&tempData->mapsData[1].height);
+    tempData->mapsData[1].data = LoadMapTiles(ASSETS"/Maps/CGame_L1_UP_Layer.csv",&tempData->mapsData[1].width,&tempData->mapsData[1].height);
     tempData->mapsData[1].size = (Rectangle){0,0,(float)tempData->mapsData[1].width * 16, (float)tempData->mapsData[1].height * 16};
     ///< Load Textures
     ///<    TileMap && EmptyTexture
-    tempData->tileMapTex = LoadTexture("assets/Tilemap/TiledMapWorld.png");
-    tempData->emptyTexture = LoadTexture("assets/Tilemap/EmptyTexture.png");
+    tempData->tileMapTex = LoadTexture(ASSETS"/Tilemap/TiledMapWorld.png");
+    tempData->emptyTexture = LoadTexture(ASSETS"/Tilemap/EmptyTexture.png");
     return tempData;
 }
 //////////////////////////////////////////////////////////////////